backend/requirements.txt
Flask>=3.0.3
Flask-PyMongo>=2.3.0
Flask-Cors>=3.1.0
gunicorn
pytest
pytest-cov

